api/camera/requirements.txt
fastapi[standard]==0.115.8
pillow==11.1.0
